+++ selfdrive/car/chrysler/radar_interface.py
@@ -5,6 +5,7 @@ from cereal import car
 from opendbc.can.parser import CANParser
 from selfdrive.car.interfaces import RadarInterfaceBase
 from selfdrive.car.chrysler.values import DBC, PRE_2019
+from common.params import Params
 
 RADAR_MSGS_C = list(range(0x2c2, 0x2d4 + 2, 2))  # c_ messages 706,...,724
 RADAR_MSGS_D = list(range(0x2a2, 0x2b4 + 2, 2))  # d_ messages
@@ -12,6 +13,9 @@ LAST_MSG = max(RADAR_MSGS_C + RADAR_MSGS_D)
 
 
 def _create_radar_can_parser(car_fingerprint):
+  if Params().get_bool("jvePilot.settings.visionOnly"):
+    return None
+  
   dbc = DBC[car_fingerprint]['radar']
   if dbc is None:
     return None