+++ RELEASES.md
@@ -8,6 +8,7 @@ Version jvePilot-0.9.0(v3.13.0)
 * LKAS button can now toggle experimental mode that slows for stop lights and stop signs.
   * Enable Experimental Mode in the dash will make the mode available to toggle using the button
   * When button is lit, experimental mode will be enabled.
+* Add option to ignore the radar and use vision only 
 * Add Fingerprint V1 for AUS Jeep Grand Cherokee Trailhawk
 * Update to 0.9.0
   * New driving model